PESDLC-1487 Transaction tests consolidation
    Moves transaction related tests to subfolder and fixes
    some type checks.
