+++ tests/rptest/transactions/tx_abort_index_test.py
