+++ tests/rptest/transactions/tx_verifier_test.py
