+++ tests/rptest/transactions/verifiers/compacted_verifier.py
